Target-driven update complete
Now we step through a list of targets from the BED file without concern
for order in our reference sequence.  This makes our code simpler and

However, doing this required some rewriting of the sequence loading
sections.  As alignments at the start and end of our target sequence
will lie outside of the bounds of the target, we must grab more sequence
than target, allowing the analysis of of the mapping of all of the reads
overlapping the target sequence.  For this iteration I'm processing
reads at the beginning and end to exactly determine the length of the
left and right overhangs; but in the future it may prove simpler (and
much faster) to select a length greater than the longest read in the
dataset but much shorter than the full reference sequence length (as was
done previously).