Untested incremental commit
This code probably doesn't even compile.  This commit serves as a
checkpoint of my progress over the past few days.

In this commit I include a heavily commented and incomplete function,
Caller::probObservedAllelesGivenGenotype, which estimates the
probability of the observation of a set of alleles given an underlying
genotype and a ploidy.  This function generalizes the mathematical
techniques found in GigaBayes.