Allele observation probabilities, Allele probabilities
This commit includes (compiled, but thus far untested) code to calculate
allele observation probabilties (probability that we have a true
observation of an allele) and allele probabilities (multinomial
probability of a set of observations given an underlying genotype).

This functionality is provided via a function that takes a set of allele
observations from an individual sample and a list of genotypes.  The
function returns a vector of probabilities ordered according to the
input genotypes.