FASTA reader fix, math simplification, optimization
Incorporats a fix to the FASTA reader resolving a problem in which FASTA
index entries were improperly sorted.

Numerous tweaks to the math; checkpoint in this sense.  The system is
now essentially generating genotypes for each individual.  Math
attempting to estimate P(all genotype choices | obserations) has been
removed, and will be reincorporated after discussions with Gabor.  My
objective is to improve the performance of this aspect to the point that
the system can massively reduce data, perhaps to a degree commensurate
with the limitations of a scripting-language based P(SNP) estimator.

I applied the -O3 compiler flags and found a large (2x) improvement in
performance.  I also noticed that nearly half the system runtime is
spent dealing with Allele creation and destruction.  In the subsequent
commit I will move most Allele manipulation functions to work on
Allele*, an avenue I elected not to go down until I had sorted through
most of the algorithmic issues.