Object recycling for Allele*
In earlier tests I found that much (up to 40%) of runtime was spent
creating and destroying Allele objects.  I implemented a free list based
recycling system.  The result is mixed; in return for a reduction of
time spent copying Alleles the system now incurs a slight dereferencing
penalty whenever it works with them.  Overall performance slightly
improved.  Now a bottleneck lies at the Allele sorting step.  As Alleles
are always manipulated in terms of their sample membership, I am going
to change the primary accounting system for them from a list of all
Alleles in alignments overlapping the current location to a map from
sample id to Allele list.