long double, json output
I've moved math to use long doubles instead of doubles.  (No significant
performance loss noted, precision is necessary for calculations when
coverage is >150 or so.)

This commit produces a json-formatted output stream.  In production this
steam will then be piped to a processing application that can output a
variety of interchange formats, including VCF and GLF.