stable release candidate
This is a 'stable' release candidate of bamBayes meant to bridge the gap
between the existing system and the one currently in development.
Certain improvements from the upcoming version have been incorporated.
Some testing has been completed as of this commit.

bamBayes now incorporates a multi-reader which is capable of opening any
number of sorted BAM files and reading from them as if they are one
sorted, and merged, BAM.  A list of files can be specified as a space or
newline-delimited list passed as a string to the --bam parameter (e.g.
--bam "one.bam two.bam three.bam").

This commit also includes extensive fixes to sample name (SM) handling.
Now assume, per existing BAM producers, that we can have many read
groups (RG's) per sample name (SM).  At runtime we generate a map of
RG->SM which is used to get sample names for each alignment in our
analysis.

In order to guarantee expected behavior, the system will raise an error
and exit if a sample specified in a sample file is not present in the
BAM file or files passed to the program.

VCF 3.4 output is supported alongside RPT output.