Priors partly implemented, utility functions unified
This is a development commit.  Several issues are halfway addressed,
but I wanted to track progress and provide a checkpoint for future work.

I have implemented the priors calculation with the exception of the
application of Ewens Sampling Formula.  I am leaving this for the
resolution of some problems with the data likelihoods calculation.  At
present the implementation is broken as it only considers the
most-likely case in which the underlying 'true' alleles are the same as
the 'observed' alleles.  The actual data likelihood requires the
estimation of a much larger space of possible underlying 'true' alleles
than is currently accounted for.  We should be considering all possible
multisets of true alleles to better approximate the probability space
occupied by the data likelihoods.

The next commit should include a completed priors calculation and also a
suitable approximation method for our data likelihoods.