'Low-memory' BamBayes
Instead of reading in entire references at runtime, which consumes
enormous amounts of memory and makes it impossible to run multiple
instances of BamBayes on a single (small) cluster node, simply load the
target regions from the reference as we process them.  This will save
many gigabytes of runtime memory when processing human datasets.