BamTools update, alleles.cpp, python data likelihood calculations
Commit checkpoints a Python data likelihood calculator which takes the
observed alleles stream from alleles.cpp and annotates each sample with
a list of genotype likelihoods given the observed alleles.  This is not
intended for production use, but rather to speed debugging and
implementation of the core algorithm.

Subsequent work will add a genotype combination calculator which follows
the algorithmic development for the Bayes/Ewens variant caller.