Completed python implementation of full algorithm
I have (mostly) completed and tested the python implementation of the
current algorithm, which features:

1) A Bayesian framework for establishing the genotypes of a population
of individuals aligned to a common reference (same as GigaBayes,
BamBayes).

2) The use of Ewens' Sampling Formula for the estimation of genotypes on
the level of the population of individuals under analysis.  It is used
to establish a suitable prior genotype combination probability.
Importantly, it allows us to establish this probability without
limitation of the ploidy (copy number) of the individual genotypes; each
individual may have a different copy number.  The concept of allele is
abstract, and thus the formula may be applied to more abstract loci than
single-base reference positions.

3) Dirichlet distribution-based priors for multinomial samples, used in
the estimation of data likelihoods (genotype likelihoods) for each
individual.  This prior probability is derived from the ratio between
the Dirichlet probability density for specific sequencing samplings of
the genotype and the maximum probability density for the specific
genotype and number of observations.  A concentration parameter allows
us to tune our confidence that the sequencing observations should follow
a perfect multinomial distribution.  As the Dirichlet distribution is
conjugate to the multinomial distribution, we are assured that our
posterior data likelihoods are properly distributed.

4) Exhaustive, complete estimation of data likelihoods by integration
across all possible non-zero probabilities which might result from
errors in reads.  (To be optimized and potentially approximated in
future revisions.)

5) Partial estimation of posterior normalizer by integration across
highest estimated probability genotype combinations.  Also, to be
optimized in later revisions, possibly by inclusion of some random
sampling behaviors across the posterior normalizer.