component drafting, redaction of Dirichlet priors
Addition of many source files which modularize the codebase.  These
modules now all compile, with the exception of Caller.o, which has many
problems and awaits further cleanup.

Data likelihood calculations are now implemented in their own source
file, as are Dirichlet likelihoods and multinomial probabilites.

Product.h and Sum.h encode templates which can be used to find the
product and sum of vectors of numbers.

A new Genotype defines genotypes to be a collection of counts of
different alleles.  (e.g. AT = {[1, A], [1, T]} and TT = {[2, T]})  This
simplifies probability calculations.  Also, added ALLELE_GENOTYPE to
define Allele objects which are solely used as components in Genotypes
for comparison to observed alleles.

I corrected an error I made in implementing the priors; I attempted to
use an idiosyncratic Dirichlet prior because I had improperly
implemented the multinomial sampling distribution previously (and had
suspected something amiss with the algorithm...  incorrectly).  I
switched this implementation back to using the multinomial sampling
probability as a scaling factor, and am writing the C++ version to
match.