math cleanup
Using the python implementation as a guide I resolved an issue in which
the genotype combination priors were incorrectly calculated.  The
current fix is insufficient.  I need to do some work on the Allele data
type's comparison operators and test its use in maps, as this was the
source of the issue.