performance enhancement, priors sampling improvement, new parameters
1) By caching data likelihoods on the basis of the number of errors
implied by the genotype against which the likelihood is conditioned, we
can eliminate a large fraction of the computation for most individuals
in low-moderate coverage.  This provides a 2x performance boost.  More
radical heuristic optimizations of this calculation are to follow.

2) I noticed that in some cases the banded approximation combined with
ewens sampling formula encouraged the propagation of errors into the
best genotype combo.  In many cases the best homozygous combination
should have much higher priors.  In order to enforce the evaluation of
this important position in the posterior probability space, I check if
the list of genotype combos includes at least one combo in which all
individuals are homozygous.  If none is included, I append the combos
list with the best possible homozygous combo according to the data
likelihoods.

3) I've included vcf output to stdout as a command line parameter.