Resolved exp underflow bug
The logsumexp function in Utility.cpp is correct, but unsafe as it
doesn't check for overflow.  This produces errors in the case of very
small probabilities, and in prior revisions alleleBayes would simply
skip positions for which this underflow occurred in too many genotype
combinations.

To resolve the issue I have implemented safe_exp(), which returns 0 in
the case of an underflowing value, and logsumexp_probs(), which
normalizes the largest probability in its input vector to 1, and scales
its input vector by the normalization amount before executing the
exponential summation using safe_exp().  In other words, it will be
assured that at least one value in the input vector is non-zero, and
logsumexp_probs() will always return something useful.

I additionally removed an inappropriate filter, BQL2, which should have
been BQL0.  I will be implementing filters in the next commit.