memory leak cleanup, trialleleic reporting, input filter fixes
Memory leaks were present in a number of locations, the most egregious
was in FastaReference::getSubSequence(), which would leak at least byte
for each reference base it returned.  Also, the system leaked the
reference allele if front-end filters chose to skip the position.  I
resolved these issues and investigated if any leaks could be attributed
to the Allele recycling system, but I have not found any which can.
Some memory appears to be un-reclaimed through the runtime of the
system; it is bound up by BamIndex.  This is relatively minor in
comparison to the other leaks though, so resolution should be able to
wait a longer discussion with Derek.

Unfortunately I was unable to commit changes which fixed the triallelic
reporting mechanisms prior to implementing the memory leak cleanup, so
those (minor) changes also fall into this commit.