fix: output for all positions, regardless of observations
Previously we didn't properly handle output for positions in which the
best genotype combination was homozygous (instead these were typically
skipped...).  This is a step in the right direction.  Also includes a
parameter to force the inclusion of all possible bases in the possible
genotypes so that the program behaves the same as bambayes on sites
which otherwise have no variation.