performance tweaks
indel mask -> vector<bool> (faster than a map); also stored as 0-size
vector in the case of no masking.

qualities; attempt to improve 'currentquality' performance, as this is
now 10% of runtime.  didn't quite work, yet.

no longer store a copy of reference sequence for each allele; this is
simply wasteful as we have a copy available in the AlleleParser.

other fixes, tweaks, updates.