bugfix: remove filtered alleles from sampleAlleles
In order to ensure that no alleles which should be filtered are
presented for analysis, we must remove filtered alleles from the running
sampleAlleles map, and requeue them for processing.  If we don't do this
then the BQL0 filter will have little or no effect, and the indel window
will also fail in many cases.  This commit resolves the issue.