extensive performance improvements
Many small optimizations.  Principally, we limit the depth of the
integration of marginal genotype probabilities to the best K genotype
combinations, specified by -K on the command line.  In a small test case
across 600+ samples, the result is that we are super-linear in time in
the order of input samples, but by a smaller factor than before (only 3x
as opposed to 8x per doubling).