version 0.2.0: performance enhancements
This commit (v0.2.0) constitutes a significant rewrite of the internal
structures used to calculate genotype combination likelihoods.  In
profiling it was revealed that some structures used in prior versions
had poor scaling characteristics as the number of samples in the
analysis increased.  To resolve this issue, setup algorithms were
adjusted to:

  1) Group alleles by type (e.g. A, C, G, T, Indels) and by sample in
AlleleParser::getNextAlleles(...).  This removed unnecessary sorting and
iteration through sets of observations.
  2) Calculate genotype combination allele frequency and data likelihood
by adjusting them only for those samples which differ from the best
genotype combination.  This prevents O(N^2) performance in establishing
the probability of each genotype combination independently.
  3) Use pointers for heavyweight data structures (e.g. GenotypeCombination
Results).
  4) Index maps of Allele*'s by the string representation of the Allele,
so map<Allele, *>'s were moved to map<string, *>.  This obviated the
overuse of operator<(Allele&, Allele&), which in turn just used
Allele.currentBase.

No algorithmic changes were required to improve performance, and testing
suggests that the results as of this version are identical to v0.1.0
except for very small differences (+/- 1) in the marginal genotype
likelihoods.  I am testing to establish the source of this change and if
I discover a significant issue an upcoming point release will include a
fix.

As of this commit, the --posterior-integration-depth, -K must be
specified or it is assumed to be unbounded, in which case every
combination which is established in the first step of the posterior
predictive calculation will be used to calculate marginals and the
posterior predictive.  Previously the default was 100.