parameterize 'banddepth'
In examining discrepancies in marginal likelihood output between this
version and 0.1.0 I added parameterization for the posterior banddepth.
In practice, with a large number of samples, it is a very bad idea to
move this away from 1, as it will induce the integration of a very large
subset of the posterior predictive space!  I thus have removed it from
the help dialog to discourage use.

Additionally, I discovered that the discrepancies in marginal likelihood
were related to a bug which I resolved in the previos version (0.1.1)
where the genotype combination generation omitted one combination.