enable calling of insertion and deletion (indel) alleles
As of this commit indel calling is tested and fully implemented.  The
implementation is incomplete in that the prior probability algorithm has
not been adjusted to the lower expected probability of indel events.

Indels are reported in VCF as multi-base alternate strings for
insertions and, for deletions, DN where N is the number of deleted
bases.  I may subsequently change this to include the deleted bases in
the REF field in the VCF output, but I need to review the ever-changing
specification.