compute P(genotype combo|allele frequency) using flexible algorithm
Previous to this commit, we used an algorithm which estimated the
probability of a given genotype combination given an allele frequency
under the assumption that all individuals were diploid.  I have updated
this calculation to use a formula which allows

In the diploid case it should produce the same results.  It will allow
us to operate on higher copy-number, but some bias may result from the
fact that this adaptation makes no distinction between heterozygotes
which have different numbers of alternate alleles (e.g. 1/1/0 and 1/0/0
are treated equivalently).  I will establish if this effect may cause
undue bias in this approximation.