enable calling of MNPs (multi-nucleotide polymorphisms)
Previously MNPs were parsed as multiple SNPs.  This is both incorrect
from the perspective of SNP detection and limited in that it prevented
the detection and genotyping of multi-nucleotide variations.  MNP
detection is enabled using the --mnps command-line flag.  They are
flagged in the VCF output by the "MNP" flag.