allow running on BAM files without specified read groups
Many users are attempting to run freebayes on BAM which do not contain
specified samples and read groups.  BWA will produce such BAM files
unless users are careful to build sample information into the sample
files during the alignment process [1].  Rather than exit and prevent
analysis in this situation, freebayes will now display several warning
messages to stderr and proceed with the analysis.

Additionally, users attempt to run freebayes on single-sample BAM files.
This defeats the major benefits of the freebayes calling algorithm, but
it should be allowed and handled gracefully.  It is now allowed, but
with a warning that '--use-reference-allele' and
'--force-reference-allele' are assumed in this case.  Without these
flags freebayes will only report heterozygous sites from the same
individual.

[1]
http://www.broadinstitute.org/gsa/wiki/index.php/BWA_patch_to_generate_read_group