alignment.GetEndPosition() is inconsistent
For reasons I haven't yet debugged, alignment.GetEndPosition() can
return the wrong position in some cases.  This has been most notable in
SOLID/bfast data, but seems to occur elsewhere (such as in the NA12878
raw bwa alignments), although I don't have that dataset on hand to test.

The error mode is a segfault caused when the system attempts to use an
allele pointer to an allele in a reclaimed registeredAlignment.  The
registeredAlignment is reclaimed too early because
alignment.GetEndPosition() returns too low a position (by 2 bases in one
case).

To resolve the issue, I obtain the end position using alignment.Position
+ alignment.AlignedBases.size().  This was the method used previously,
but I switched to alignment.GetEndPosition() as it seemed more
straightforward and correct.