insertion and deletion quality
Give insertions and deletions quality which corresponds to either the
sum of quality (product) of inserted sequence, or the flanking sequence
of a deletion of length corresponding to the length of the deletion.
The logic is that these events involve multiple bases of observation,
and thus their qualities should reflect the greater amount of
information included in their observations.  Additionally, the use of
flanking sequence for deletions is important as it is the quality of the
sequence of the read flanking the deletion which determines if it is
placed correctly by the aligner.

In tests in conjunction with BAQ adjustment (via samtools) it appears
that these changes generate insertion to deletion ratios much closer to
1:1 than previous methods in freebayes.