improve indel quality estimation
Add --report-all-alternates, which outputs one line of VCF per alternate
allele at a site, instead of just outputting the first best alternate.

Adjust insertion and deletion quality estimation:

Deletion quality is proportional to the number of bases in the deletion
flanking the position of the deletion, scaled by the ratio between the
number of bases used in the quality estimation and the number of bases
in the deletion.  This scaling avoids stair-step behavior in the quality
function.

Insertion quality derives from the quality of the bases within the
insertion and the flanking right and left base.  This score is also
scaled by the ratio between the length of the insertion event and the
length of the bases which provide quality information for the event.

The use of flanking bases smooths the quality values of single base
insertion events.  This use is well-justified, as flanking
reference-matching bases are the most important determinants of indel
position or read placement.  Additionally, this technique, allows us to
optionally use base alignment quality adjustment on the input BAM stream
without introducing artifactual scores for short indels.