bamleftalign, bamfiltertech, filters for indel detection
bamleftalign is a read-independent indel realignment system which reads
BAM on stdin and writes BAM on stdout, and is designed for use with
freebayes.  It iteratively shifts indels as far left as they can go
without introducing mismatches.  This assists in the detection of indels
by homogenizing the placement of identical events which may have been
aligned differently due to characteristcs of alignment.  Left alignment
is also standard for the 1000 Genomes project (see the 2010 07 indel
release), and allows indel callsets to be compared easily.  In practice,
left realignment is very efficient, and realigned output can be streamed
into freebayes or other programs, such as samtools base alignment
quality adjustment.

bamfiltertech is a BAM stream filter which can be used to restrict
analysis to only reads derived from a certain set of technologies listed
on the command line (e.g.  illumina and solid, but not 454).  This is
important for indel detection as some technologies have many artifactual
indels, and they may be mixed with other reads in large project BAM
files.