per-sample, per-region copy number variation
By providing a map of copy number variation across individuals
(specifically, a BED file of the format <seq>, <start>, <end>, <sample>,
<copy number>), it is now possible to configure variant copy number
across the entire population.  (see --cnv-map).

This allows us to, for instance, call variants on the human X chromosome
without segregating our population into male and female cohorts.

The functionality is untested as of this commit, but the commit is
provided as a checkpoint.