extension of left alignment to all repeat lengths and shift sizes
The algorithm is adjusted to step up the attempted left shift increment
until it is the size of the indel.  This ensures that repeats smaller
than the length of the indel event are caught by the left realignment.
They were not previously.  This extends the classes of left shifting
which the algorithm handles.