version 0.4.0, optimizations, code restructuring, and output changes
This commit contains a number of changes:

1) Refactoring of codebase to move Sample into its own source file.
Removal of circular dependencies between Sample, Genotype, and Allele.

2) Cleanup of bottleneck in Sample::sortAlleles, changed to specifically
handle reference alleles, as this was the only case that occurs in
practice.  Removed unnecessary second call to function.  Renamed to
sortReferenceAlleles to reflect proper usage.

3) Implementation of data likelihood based step filter, which removes
genotype combinations which would include a non-maximum genotype from a
sample which is more than some defined log(step) away from the maximum
likelihood genotype for that sample.  This reduces the number of chains
which must be operated on, improving runtime, but removing chains which
might have very little effect on the results.

4) Removal of option to use the factorial data likelihood calculation of
BamBayes/GigaBayes.

5) Modification to the VCF output stream, which adds separate fields for
each paired tag that was previously encoded as a <reference>,<alternate>
pair.  Addition of often-requested fields, specifically minor allele
frequency, overall reference and alternate observation counts.  These
changes should make it easier to implement multi-alleleic reporting.

6) Copy of multiset permutations as a functor, class
MultisetPermutations.  This was done to explore the performance
characteristics of using a functor to generate each permutation on-line
instead of returning the entire vector<vector<int> > of indecies for the
genotype combination chains.  In practice, this slowed down runtime
significantly, on the order of 5%.  The effect could be an artifact of
the implementation.