set supporting quality filters to 0 by default
A bug in a previous commit unset the min-supporting* quality filters.
The last commit reinstated them.

In practice, the default stringency of 30/40 for quality and mapping
quality appears to be too high, provided we are already filtering low
quality alleles and low mapping quality reads.  This commit sets the
filter settings to 0/0, so that behavior does not change, but it is now
configurable.