left-right base balance
Characteristic error in seeded local alignment often generates an
imbalance between the number of bases to the left and right of putative
variant alleles.  Tracking this imbalance requires accounting the number
of bases to the left and right of each "Allele" within the sequence read
it is derived from.  Now we can report the left-right balance (LRB) for
each alternate allele in the output.