Hoeffding's inequality estimates for "balance" variables
The balance metrics, such as SB (strand bias) and AB (allele balance)
are simple ratios, and as such cannot capture the likelihood that the
bias is as great as observed given an expected ratio of 0.5.

A more appropriate metric for filtering is provided by Hoeffding's
inequality, which can be used to obtain the upper bound for the
probability of deviation from expected as great as that observed in some
data.

Hoeffding's inequality may be computed in constant time, and is thus
preferable to the corresponding exact metric for binomial distributions,
which requires O(N) time to calculate, where N scales with the deviation
in the observations from expected.

This commit adds deviation probability estimates for strand bias, allele
balance, and read placement balance.  They are provide as Phred-scaled
floats, to facilitate parsing and rapid interpretation.