region spec generation from fasta index file
This script can be used to make region specifiers which can in turn be
used in a map/reduce operation to map freebayes across a whole genome.
One would launch a single job per region specifier, changing no other
parameters, and output to a different file, then merge the files.