performance enhancement for large numbers of samples
In genotype combination generation, avoid use of a very expensive copy
constructor by building vectors of pointers to data likelihood objects.
This normalizes time complexity to mostly O(N) in the order of the
input, unless you want to estimate the marginal probabilities of
genotypes for each sample (presently done by default, but disabled via
the --no-marginals flag).

For large numbers of samples, the performance improvement offered by
this commit is very significant.