expectation maximization for maximum a posteriori estimate
This commit adds an EM/descent method for finding the maximum a
posteriori estimate for the best genotyping across all individuals.

The algorithm iterates by finding the most likely genotyping across all
individuals, and then using that to seed a local search for a
more-likely, similar, genotyping using the same search parameters used
for the standard (brute force) posterior search method.  When an
iteration returns the same estimate for the local maximum, we assume the
results have converged.  The rest of the calling algorithm works the
same way.