marginals, sequencing technology output, algorithm modification
Marginals:  Use marginal probabilities established across all sampled
population genotypings to estimate genotypes for individuals.  Ignore
the homozygous genotype combos (for now), as the structure of the priors
will suppress the output of monomorphic sites.

Sequencing technology: Output the fraction of observations from each
sequencing technology in the input supporting a given alternate in each
line of VCF.  This will allow differential filtering of sites which are
only supported by one technology and which are characteristic errors of
that technology (e.g.  short homopolymer indels in 454).

Algorithmic modification: remove the secondary scaling term (the
multinomial coefficient of heterozygous genotypes) from the genotype
combination probability given allele frequency component of the priors.
It seems that this component only makes sense if we are evaluating all
possible phased genotype combinations, which we are not.  Most
importantly, it appears that results improve markedly when removing this
component.

This is a provisional, checkpoint-style commit.  Version 0.7.0 will
combine these changes with genotype refinement by HWE estimation.  (For
convenience, many changes exist in this commit as commented-out blocks,
but these will be removed or parameterized in subsequent commits.)