version 0.8, input of variants via VCF
Users may specify an input VCF file in bgzip, tabix-indexed format to
use to call variants.  This file should have sample-specific genotypes
and, ideally, observation counts for the alternate alleles in called
genotypes.  This information is then used as additional observations
with which to co-compute genotypes of samples in the input BAM data.

The output then includes only information about the samples in the input
BAM, although they have been modeled alongside variants and genotypes
reported in the input VCF.  The depth and QUAL reported in the output
reflect the entire set of samples from both the VCF and the BAM, but
other fields only reflect the BAM-input samples.

Methods supporting a sites-only input will be implemented in subsequent
revisions.