enable input of alleles in VCF without genotypes (version 0.8.1)
This commit expands on the previous work which enabled the input of
genotype observations in VCF format by allowing the use of VCF files
without assigned genotypes.  Now, every allele provided in the input VCF
is evaluated in the context of the current data.  Output is provided for
each of the input alleles.

For instance, this allows the use of an input VCF file for re-evaluating
genotype likelihoods within a different alignment set.

It also can be used to improve non-reference allele detection, as the
input alleles are not put through any of the input filters which are
applied to the input alignments (BAM).