add --only-use-input-alleles, restricts alleles to those in input VCF
This flag restricts evaluated alleles and ouput alleles to those in the
input VCF.  The output will include information about *all* the alleles
in the input.

This enables the calculation of genotype likelihoods (or any other
metrics provided in the output VCF) for an arbitrary set of alleles and
sites.