don't use reference allele as a sample unless directed to
Except in the case of many samples, the use of the reference allele as a
sample can overwhelm evidence for non-reference variation, causing some
samples to be incorrectly called as reference.

This effect particularly pronounced in haploid organisms.  Two users
provided test cases in which two haploid organisms were incorrectly
called when the reference allele was considered a sample.

This commit inverts the behavior of the "-Z" flag, and changes the long
form name from --ignore-reference-allele to --use-reference-allele.