allow complex alleles to have embedded matching sequence (v0.9.0)
With this change, complex alleles are generated for cases where two
small variants in the same read occur at most --max-complex-gap bases
apart (3bp, by default).  This allows for the detection of MNPs
(multi-nucleotide polymorphisms) with embedded matching bases.  The
behavior can be disabled by setting --max-complex-gap 0.

Also, this commit adds a new tag to the VCF output, CIGAR, which
provides the CIGAR strings of the variants, allowing for post-hoc
filtering of certain classes of complex variants.