resolve haploid genotyping bug
Due to a recent change in the way that the reference allele is handled,
in some cases it was possible that the best genotype combination was not
evaluated when calculating marginal genotype likelihoods.  As a result,
genotypes were frequenty mis-called in the case of two haploid samples.
This resolves the bug by ensuring that the best genotype combination is
added to the set of combinations that are evaluated.