performance improvements
1) Introduce the concept of a null allele.

This is used in the place of all the other potential alleles at the site
when calculating genotype likelihoods for a given sample.  If the sample
does not have any observations for a given alternate allele, we just
ignore it.  The likelihoods for such genotypes are then provided by
matching the genotype to one in which the missing allele is replaced by
a null allele.

The benefit of this is that we dramatically reduce the number of
potential genotype combinations which we have to evaluate when searching
the posterior space for the maximum likelihood solution.  This is done
without any serious change to the algorithm design, and allows marginals
to be calculated without issue.

2) Cache binomial calculations.  This provides a 15% speedup when using
--binomial-obs-priors.

3) Don't store intermediate genotype combination results.  Doing so
causes severe memory blowups.  (I'm also considering changing the
GenotypeCombo to a vector<short>, and including some kind of genotype
ptr -> short int mapping for the combo.)