exclude alleles with no reference sequence
These are generated in the process of haplotype construction.  Call them
chaff; the alleles which they have been carved out of could not pass
relatively minimal filter cutoffs, and as such are carved up. It's very
unlikely that they are significant, and reconstructing them properly
would require a lot of code adjustment and probably would not result in
better performance.