population subdivisions
These changes allow the subdivision of the input samples into
sub-populations.  The sub-populations are assumed to be inbreeding,
selectively neutral, random samples.  Provided this, the model is
evaluated for each population independently.

The sub-population model assumes independence among the populations.  At
present, mutual information is shared between populations only in the
sense that alleles and genotypes evaluated for one population are
evaluated for all.

Populations may be specified using a file mapping sample names to
populations.  The command-line flag is --populations.