homogenize alternate alleles at haplotype loci
Depending on sequence context, a complex allele which is 1M5D1M1D is
potentially the same as a deletion allele 2M6D.  This equivalence can be
established by comparing the alternate sequences for a given
reference-relative haplotype.  The most-commonly-observed alignment is
used to adjust the cigars for identical but differentially described
alternate alleles.