haplotype basis alleles
By specifying a set of haplotype basis alleles, phasing information can
be established in long reads even in the presence of high error rates.
The haplotype basis allele input is used to select the alleles which
will be phased.  Other alleleic primitives will be ignored by
replacement with the reference allele.