remove default mapping quality and base quality restrictions
The mis-estimation of mapping quality causes a lot of problems for
users.  Largely, these issues can be resolved by removing the default
input filters in freebayes.

A better method of incorporating mapping quality into the analysis is
generate genotype likelihoods using the minimum of base quality and
mapping quality.  This can be enabled by providing the
--use-mapping-quality flag on the command line.