build haplotypes across repeats (version 0.9.7)
When an indel is based on underlying repeat structure, record the right
boundary of the repeat in the reference (technically, the first base
past the repeat) in the indel's Allele structure.  When building
haplotype alleles during genotyping, assemble across the repeat,
requiring, for instance, reference-matching reads to cover the entire
repeat sequence.