set default input filters (-C 2 -F 0.2)
In testing, these input filters on the minimum support for a given
allele have been found to provide a very good balance between
sensitivity and specificity, reducing the need for users to place
complex filters on their VCF output.

We use them by default in our work in the 1000 Genomes project
low-coverage (4-6x) data.

However, They may not be ideal in polyploid or pooled systems or
low-frequency somatic variant detection, so users working in such
contexts should set them to a level appropriate for their needs.