turn off genotype qualities by default
Genotype qualities (reported as GQ in the output) are marginal
likelihoods of the specific genotype for a specific sample given the
data and Bayesian model.  They may be helpful for filtering or assessing
genotyping accuracy, but they take a lot of time to compute because the
current method for estimating them is O(N^2) in the number of samples.
For more than 10 low-coverage samples, GQ estimation becomes the
dominant use of compute time.  For 1000 samples, GQ estimation is 90% of
compute time.

Prior to this commit it was possible to disable them using
--no-marginals.  I've removed this crypticly-named parameter, set GQ
estimation off by default, and added --genotype-qualities, which turns
them back on.

Users who wish to use fill out the GQ field (old behavior) must provide
--genotype-qualities.