balanced observations for indels (indels)
For insertions (currently), require that the reference observations at the
haplotype containing the insertion flank the insertion by a combined number of
bases that is the same as the length of the insertion.

This normalizes the likelihood calculations for insertions.

Normalization is (currently) provided implicitly for deletions in that
reference bias prevents the mapping of longer deletions.  This effect is
stronger than the sum of bases effect driving bias problems for insertions.